+++ DatadogRUM/Sources/Instrumentation/RUMInstrumentation.swift
@@ -27,9 +27,9 @@ internal final class RUMInstrumentation: RUMCommandPublisher {
     /// Swizzles `UIApplication` for intercepting `UIEvents` passed to the app.
     /// It is `nil` (no swizzling) if RUM Action instrumentaiton is not enabled.
     let uiApplicationSwizzler: UIApplicationSwizzler?
-    /// Receives interceptions from `UIApplicationSwizzler`.
-    /// It is `nil` if RUM Action instrumentaiton is not enabled.
-    let actionsHandler: UIEventHandler?
+    /// Receives interceptions from `UIApplicationSwizzler` and from SwiftUI instrumentation.
+    /// It is non-optional as we can't know if SwiftUI instrumentation will be used or not.
+    let actionsHandler: RUMActionsHandling
 
     /// Instruments RUM Long Tasks. It is `nil` if long tasks tracking is not enabled.
     let longTasks: LongTaskObserver?
@@ -61,40 +61,40 @@ internal final class RUMInstrumentation: RUMCommandPublisher {
         // Always create views handler (we can't know if it will be used by SwiftUI instrumentation)
         // and only swizzle `UIViewController` if UIKit instrumentation is configured:
         let viewsHandler = RUMViewsHandler(dateProvider: dateProvider, predicate: uiKitRUMViewsPredicate)
-        var viewControllerSwizzler: UIViewControllerSwizzler? = nil
+        let viewControllerSwizzler: UIViewControllerSwizzler? = {
+            do {
+                if uiKitRUMViewsPredicate != nil {
+                    return try UIViewControllerSwizzler(handler: viewsHandler)
+                }
+            } catch {
+                consolePrint(
+                    "🔥 Datadog SDK error: UIKit RUM Views tracking can't be enabled due to error: \(error)",
+                    .error
+                )
+            }
+            return nil
+        }()
 
-        // Create actions handler and `UIApplicationSwizzler` only if UIKit instrumentation is configured:
-        var actionsHandler: UIKitRUMUserActionsHandler? = nil
-        var uiApplicationSwizzler: UIApplicationSwizzler? = nil
+        // Always create actions handler (we can't know if it will be used by SwiftUI instrumentation)
+        // and only swizzle `UIApplicationSwizzler` if UIKit instrumentation is configured:
+        let actionsHandler = RUMActionsHandler(dateProvider: dateProvider, predicate: uiKitRUMActionsPredicate)
+        let uiApplicationSwizzler: UIApplicationSwizzler? = {
+            do {
+                if uiKitRUMActionsPredicate != nil {
+                    return try UIApplicationSwizzler(handler: actionsHandler)
+                }
+            } catch {
+                consolePrint(
+                    "🔥 Datadog SDK error: RUM Actions tracking can't be enabled due to error: \(error)",
+                    .error
+                )
+            }
+            return nil
+        }()
 
         // Create long tasks and app hang observers only if configured:
         var longTasks: LongTaskObserver? = nil
         var appHangs: AppHangsMonitor? = nil
-
-        do {
-            if uiKitRUMViewsPredicate != nil {
-                // UIKit views instrumentation is enabled, so install the swizzler:
-                viewControllerSwizzler = try UIViewControllerSwizzler(handler: viewsHandler)
-            }
-        } catch {
-            consolePrint(
-                "🔥 Datadog SDK error: UIKit RUM Views tracking can't be enabled due to error: \(error)",
-                .error
-            )
-        }
-
-        do {
-            if let predicate = uiKitRUMActionsPredicate {
-                let handler = UIKitRUMUserActionsHandler(dateProvider: dateProvider, predicate: predicate)
-                actionsHandler = handler
-                uiApplicationSwizzler = try UIApplicationSwizzler(handler: handler)
-            }
-        } catch {
-            consolePrint(
-                "🔥 Datadog SDK error: RUM Actions tracking can't be enabled due to error: \(error)",
-                .error
-            )
-        }
 
         if let longTaskThreshold = longTaskThreshold {
             if longTaskThreshold > Constants.minLongTaskThreshold {
@@ -150,7 +150,7 @@ internal final class RUMInstrumentation: RUMCommandPublisher {
 
     func publish(to subscriber: RUMCommandSubscriber) {
         viewsHandler.publish(to: subscriber)
-        actionsHandler?.publish(to: subscriber)
+        actionsHandler.publish(to: subscriber)
         longTasks?.publish(to: subscriber)
         appHangs?.nonFatalHangsHandler.publish(to: subscriber)
         memoryWarningMonitor?.reporter.publish(to: subscriber)