+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Unreleased
 
+# 2.19.0 / 28-10-2024
+
+- [FEATURE] Add Privacy Overrides in Session Replay. See [#2088][]
+- [IMPROVEMENT] Add ObjC API for the internal logging/telemetry. See [#2073][]
+
 # 2.18.0 / 25-09-2024
 - [IMPROVEMENT] Add overwrite required (breaking) param to addViewLoadingTime & usage telemetry. See [#2040][]
 - [FEATURE] Prevent "show password" features from revealing sensitive texts in Session Replay. See [#2050][]
@@ -774,6 +779,8 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#2043]: https://github.com/DataDog/dd-sdk-ios/pull/2043
 [#2040]: https://github.com/DataDog/dd-sdk-ios/pull/2040
 [#2050]: https://github.com/DataDog/dd-sdk-ios/pull/2050
+[#2073]: https://github.com/DataDog/dd-sdk-ios/pull/2073
+[#2088]: https://github.com/DataDog/dd-sdk-ios/pull/2088
 [@00fa9a]: https://github.com/00FA9A
 [@britton-earnin]: https://github.com/Britton-Earnin
 [@hengyu]: https://github.com/Hengyu